Perform wheel build in a unique directory (#22931)
Modify wheel builds to create a unique build directory for each build in
a unique temporary directory. This has two effects; first, builds no
longer rely on `/opt` being writable; and second, macOS builds (which
cannot be done inside a container) are better isolated.

Additionally, some directory names are changed to improve naming
consistency.

For the time being, tests are created in `/tmp/drake-wheel-test`.

+++ tools/wheel/wheel_builder/macos.py
@@ -1,21 +1,18 @@
 # This file contains macOS-specific logic used to build the PyPI wheels. See
 # build-wheels for the user interface.
 
-import atexit
 import glob
 import os
 import platform
 import shutil
 import subprocess
+import tempfile
 
 from .common import create_snopt_tgz, die, gripe, wheel_name
 from .common import build_root, resource_root, wheel_root, wheelhouse
 from .common import test_root, find_tests
 
 from .macos_types import PythonTarget
-
-# Artifacts that need to be cleaned up. DO NOT MODIFY outside of this file.
-_files_to_remove = []
 
 # This is the complete set of defined targets (i.e. potential wheels). By
 # default, all targets are built, but the user may down-select from this set.
@@ -27,18 +24,6 @@ python_targets = (
     PythonTarget(3, 12),
     PythonTarget(3, 13),
 )
-
-
-@atexit.register
-def _cleanup():
-    """
-    Removes temporary artifacts on exit.
-    """
-    for f in _files_to_remove:
-        try:
-            os.unlink(f)
-        except FileNotFoundError:
-            gripe(f'Warning: failed to remove \'{f}\'?')
 
 
 def _find_wheel(path, version, python_target):
@@ -115,9 +100,6 @@ def build(options):
         die('Nothing to do! (Python version selection '
             'resulted in an empty set of wheels)')
 
-    # Set up build environment.
-    os.makedirs(build_root, exist_ok=True)
-
     # Sanitize the build/test environment.
     environment = os.environ.copy()
     environment.pop('PYTHONPATH')
@@ -145,21 +127,19 @@ def build(options):
     # Inject the build version into the environment.
     environment['DRAKE_VERSION'] = options.version
 
+    # Create some scratch space for temporary storage.
+    scratch_root = os.path.expanduser('~/.drake-wheel-build')
+    os.makedirs(scratch_root, exist_ok=True)
+    scratch_dir = tempfile.TemporaryDirectory(
+        dir=scratch_root, prefix='scratch-')
+
     # Create the snopt source archive (and pass along as an environment var).
-    snopt_tgz = os.path.join(resource_root, 'image', 'snopt.tar.gz')
+    snopt_tgz = os.path.join(scratch_dir.name, 'snopt.tar.gz')
     environment['SNOPT_PATH'] = snopt_tgz
-    _files_to_remove.append(snopt_tgz)
     create_snopt_tgz(snopt_path=options.snopt_path, output=snopt_tgz)
 
     # Build the wheel(s).
     for python_target in targets_to_build:
-        if os.path.islink(wheel_root):
-            os.unlink(wheel_root)
-
-        wheel_versioned_root = os.path.join(
-            build_root, f'python{python_target.version}', 'wheel')
-        os.symlink(wheel_versioned_root, wheel_root)
-
         build_script = os.path.join(resource_root, 'macos', 'build-wheel.sh')
         build_command = ['bash', build_script]
         build_command.append(options.version)
@@ -177,13 +157,15 @@ def build(options):
         if options.extract:
             shutil.copy2(wheel, options.output_dir)
 
-        os.unlink(wheel_root)
+        if not options.keep_build:
+            shutil.rmtree(os.path.realpath(build_root))
+            os.unlink(build_root)
 
-    if not options.keep_build:
-        shutil.rmtree('/opt/drake-dist')
-        shutil.rmtree(build_root)
-        if options.test:
-            shutil.rmtree(test_root)
+    # TODO(Aiden2244) move this statement in the for loop after the
+    # `_test_wheel` statement and refactor `provision-test-python.sh` as
+    # necessary once tests are moved to a unique directory.
+    if options.test and not options.keep_build:
+        shutil.rmtree(test_root)
 
 
 def add_build_arguments(parser):
